requirements.txt
django
django-extensions
djangorestframework
gunicorn
psycopg2-binary      
python-dotenv    
django-cors-headers 
requests       
drf-spectacular 

mediapipe
opencv-python-headless 
google-cloud-aiplatform
numpy
pillow         
langchain>=0.3.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
pandas>=2.0.0
markdown>=3.5.0
groq>=0.9.0
python-dotenv>=1.0.0
